+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 * Make as-needed dropping and coalescing much faster: Don't keep scanning through the whole layer looking for another small feature.
 
+## 2.15.2
+
+* Change hash function to fnv1a
+
+## 2.15.1
+
+* Correct mbtiles inserts to use text instead of blob
+* Add an internal data structure to represent tileset metadata
+
 ## 2.15.0
 
 * Generate label points in a more straightforward checkerboard, and fewer of them at high zoom levels.